refactor(table): order 에 관한 의존성 역전을 통해 package 순환 제거
