+++ src/main/java/kitchenpos/order/application/OrderValidator.java
@@ -4,4 +4,6 @@ import kitchenpos.table.domain.TableGroup;
 
 public interface OrderValidator {
 	void validateUngroup(TableGroup tableGroup);
+
+	void validateChangeEmpty(Long orderTableId);
 }
